# coding-challenge/gen/types.gen.go
// Code generated by oapi-codegen -- DO NOT EDIT.
// Source:  api/openapi.yaml
// Config:  oapi-codegen.yaml
// Command: oapi-codegen --config=oapi-codegen.yaml api/openapi.yaml
//
// To regenerate: make generate
//
//go:generate oapi-codegen --config=../oapi-codegen.yaml ../api/openapi.yaml

package gen

import "time"

// -----------------------------------------------------------------------------
// Enums
// -----------------------------------------------------------------------------

// TransformStatus is the lifecycle state of a registered VRL transform.
//
//   - COMPILING — background compilation in progress; not yet usable.
//   - ACTIVE    — compiled successfully; ready to attach to router instances.
//   - FAILED    — compilation failed; see Transform.ErrorMessage for details.
type TransformStatus string

const (
	TransformStatusCOMPILING TransformStatus = "COMPILING"
	TransformStatusACTIVE   TransformStatus = "ACTIVE"
	TransformStatusFAILED   TransformStatus = "FAILED"
)

// -----------------------------------------------------------------------------
// Request / Response bodies
// -----------------------------------------------------------------------------

// CreateTransformRequest is the JSON body for POST /v1/transforms.
// Both Name and Vrl are required; Description is optional.
type CreateTransformRequest struct {
	// Name is a unique, lowercase-kebab-case identifier for the transform.
	// Pattern: ^[a-z0-9][a-z0-9\-]*[a-z0-9]$, max 128 chars.
	Name string `json:"name"`

	// Description is an optional human-readable explanation of what the
	// transform does. Max 512 chars.
	Description *string `json:"description,omitempty"`

	// Vrl is the Vector Remap Language program source code applied to each
	// log record before fan-out.
	Vrl string `json:"vrl"`
}

// Transform is the canonical resource returned by the API.
// It mirrors the OpenAPI Transform schema exactly so that generated and
// handwritten code share one type definition.
type Transform struct {
	// TransformId is the stable, UUID-format identifier assigned at creation.
	TransformId string `json:"transformId"`

	// Name is the customer-supplied, unique name.
	Name string `json:"name"`

	// Description is the optional human-readable description.
	Description *string `json:"description,omitempty"`

	// Vrl is the VRL program source.
	Vrl string `json:"vrl"`

	// Status is the current compilation lifecycle state.
	Status TransformStatus `json:"status"`

	// ErrorMessage is populated when Status == FAILED.
	ErrorMessage *string `json:"errorMessage,omitempty"`

	// CreatedAt is the RFC3339 timestamp of when the transform was registered.
	CreatedAt time.Time `json:"createdAt"`

	// CompiledAt is the RFC3339 timestamp of when compilation completed.
	// Null until Status transitions to ACTIVE.
	CompiledAt *time.Time `json:"compiledAt,omitempty"`
}

// TransformList is the response envelope for GET /v1/transforms.
type TransformList struct {
	Items []Transform `json:"items"`
	Total int         `json:"total"`
}

// APIError is the standard error envelope returned for all 4xx responses.
type APIError struct {
	Code    int    `json:"code"`
	Message string `json:"message"`
}

// -----------------------------------------------------------------------------
// Query parameters (generated from OpenAPI parameters section)
// -----------------------------------------------------------------------------

// ListTransformsParams holds the optional query parameters for GET /v1/transforms.
type ListTransformsParams struct {
	// Status filters results to transforms in the given lifecycle state.
	Status *TransformStatus `form:"status,omitempty" json:"status,omitempty"`

	// Name filters results to transforms whose name contains this substring
	// (case-insensitive).
	Name *string `form:"name,omitempty" json:"name,omitempty"`
}